requirements.txt
fastapi>=0.101.0
uvicorn>=0.23.2
sqlalchemy>=2.0.19
pydantic>=2.1.1
python-dotenv>=1.0.0
semantic-kernel>=0.4.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
prometheus-client>=0.17.1
pytest>=7.4.0
httpx>=0.24.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
asyncio>=3.4.3 